requirements.txt
django
python-decouple
django-debug-toolbar
django-crispy-forms
django-ckeditor
Pillow